+++ scripts/carride/gui.gd
@@ -2,7 +2,7 @@ extends Control
 
 signal touchscreen_move(direction: bool)
 
-const MINIMUM_SWIPE_DRAG = 60
+const MINIMUM_SWIPE_DRAG = 120
 
 @export var label_time: Label
 @export var label_milk: Label
@@ -12,22 +12,24 @@ const MINIMUM_SWIPE_DRAG = 60
 @export var powerup_holder: Sprite2D
 @export var powerup_ending_timer: Timer
 @export var powerup_ending_player: AudioStreamPlayer
+@export var indicators: Node2D
 
 var diff_time: int = 0
-var swipe_start: Vector2
+var swipe_start: Vector2 = Vector2.INF
+var swipe_lock := false
 var stop_processing := false
 var gauge_speed: float = -36
 # gdlint:disable=duplicated-load
 var powerup_textures = [
 	null,
 	load("res://sprites/PowerUp_SlowMotion.png"),
-	null,
+	load("res://sprites/PowerUp_Semaglutide.png"),
 	load("res://sprites/PowerUp_MilkyWay.png"),
 ]
 # gdlint:enable=duplicated-load
 var powerup_ending_tween
 
-@onready var speedometer: Label = $SpeedContainer/LabelSpeed
+@onready var speedometer: Label = $Speedometer/SpeedContainer/LabelSpeed
 @onready var score_label := $PanelContainer/MarginContainer/VBoxContainer/ScoreLabel
 @onready var game_over_panel := $PanelContainer
 @onready var play_again_button := $PanelContainer/MarginContainer/VBoxContainer/PlayAgainBtn
@@ -43,6 +45,9 @@ func _ready() -> void:
 		_enable_vbuttons(true)
 	if owner.version_2:
 		powerup_ending_timer.timeout.connect(_blink_powerup)
+	else:
+		$LivesBorder.hide()
+		$PowerupBorder.hide()
 	update_points()
 
 
@@ -69,16 +74,19 @@ func game_over() -> void:
 	game_over_panel.show()
 
 
+func direction_indicators(direction_right: bool, stop: bool) -> void:
+	var indicator = indicators.get_node("Right") \
+		if direction_right else \
+		indicators.get_node("Left")
+
+	if not stop:
+		indicator.start_blinking()
+	else:
+		indicator.stop_blinking()
+
+
 func show_powerup(powerup: PowerupClass.Powerups) -> void:
-	match(powerup):
-		PowerupClass.Powerups.RAM:
-			pass
-		PowerupClass.Powerups.SLOWMOTION:
-			powerup_holder.texture = powerup_textures[1]
-		PowerupClass.Powerups.SEMAGLUTIDE:
-			pass
-		PowerupClass.Powerups.MILKYWAY:
-			powerup_holder.texture = powerup_textures[3]
+	powerup_holder.texture = powerup_textures[powerup]
 	powerup_holder.show()
 
 
@@ -131,7 +139,8 @@ func hide_one_live() -> void:
 				0.5
 		)
 		tween.set_loops()
-
+		indicators.get_node("Service").light_on()
+		indicators.get_node("CheckEngine").light_on()
 
 
 func _gui_input(event: InputEvent) -> void:
@@ -142,15 +151,30 @@ func _gui_input(event: InputEvent) -> void:
 					touchscreen_move.emit(false)
 				else:
 					touchscreen_move.emit(true)
+		# if event.is_released():
+		# 	swipe_lock = false
+	# elif event is InputEventScreenDrag:
 		elif SettingsBus.touchscreen_control == SettingsBus.TouchscreenControlMode.SWIPE:
-			if event.pressed and event.index == 0:
+			# if event.index == 0 and not swipe_lock:
+			# 	_new_calculate_swipe(event.velocity)
+			# 	print(event.velocity)
+			if swipe_start == Vector2.INF and event.index == 0:
 				swipe_start = event.get_position()
 			elif event.index == 0:
 				_calculate_swipe(event.get_position())
 
 
+func _new_calculate_swipe(velocity: Vector2) -> void:
+	if abs(velocity.y) > MINIMUM_SWIPE_DRAG:
+		if velocity.y > 0:
+			touchscreen_move.emit(false)
+		else:
+			touchscreen_move.emit(true)
+		swipe_lock = true
+
+
 func _calculate_swipe(swipe_end: Vector2) -> void:
-	if swipe_start == null:
+	if swipe_start == Vector2.INF:
 		return
 	var swipe = swipe_end - swipe_start
 	if abs(swipe.y) > MINIMUM_SWIPE_DRAG:
@@ -158,6 +182,7 @@ func _calculate_swipe(swipe_end: Vector2) -> void:
 			touchscreen_move.emit(false)
 		else:
 			touchscreen_move.emit(true)
+		swipe_start = Vector2.INF
 
 
 func _enable_vbuttons(yes: bool) -> void:
@@ -169,9 +194,9 @@ func _enable_vbuttons(yes: bool) -> void:
 
 func _cr_speedometer_value(yes: bool) -> void:
 	if yes:
-		$SpeedContainer.show()
+		$Speedometer/SpeedContainer.show()
 	else:
-		$SpeedContainer.hide()
+		$Speedometer/SpeedContainer.hide()
 
 
 func _on_play_again_btn_pressed() -> void: